# Improve server side and client side metrics collection