After:

```
 /opt/tritonserver/bin/perf_analyzer -m identity_fp32 --shape INPUT0:1 -t 32 -f my_file.csv
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 32
  Client:
    Request count: 30
    Throughput: 1.66619 infer/sec
    Avg latency: 17695686 usec (standard deviation 0 usec)
    p50 latency: 19260511 usec
    p90 latency: 19261207 usec
    p95 latency: 19261332 usec
    p99 latency: 19261533 usec
    Avg HTTP time: 17695644 usec (send/recv 2037 usec + response wait 17693607 usec)
  Server:
    Inference count: 30
    Execution count: 30
    Successful request count: 30
    Avg request latency: 17691933 usec (overhead 99 usec + queue 17090118 usec + compute input 134 usec + compute infer 601361 usec + compute output 220 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 32, throughput: 1.66619 infer/sec, latency 17695686 usec
```

Before:

```
/opt/tritonserver/bin/perf_analyzer -m identity_fp32 --shape INPUT0:1 -t 32 -f my_file.csv
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 32
  Client:
    Request count: 30
    Throughput: 1.66587 infer/sec
    Avg latency: 17696562 usec (standard deviation 0 usec)
    p50 latency: 19260674 usec
    p90 latency: 19261328 usec
    p95 latency: 19261484 usec
    p99 latency: 19262457 usec
    Avg HTTP time: 17696519 usec (send/recv 2781 usec + response wait 17693738 usec)
  Server:
    Inference count: 30
    Execution count: 30
    Successful request count: 30
    Avg request latency: 17691667 usec (overhead 106 usec + queue 17089849 usec + compute input 136 usec + compute infer 601355 usec + compute output 220 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 32, throughput: 1.66587 infer/sec, latency 17696562 usec
```

The report doesn't look very different and I was not able to find a case where this leads to the warning being printed. Looks like all the warning I observed in the CI were related to the C-API. There is a known issue with the C-API that the queueing time is incorrect.

Having said that, the standard deviation of zero looks strange. I have filed a follow up ticket for it.