@Tabrizian 2022-07-05
>I wonder why is this change needed? From my understanding, as the windows are one after the other, we are making this change to avoid repetitive stats calls.

Previously the server stats call looked like below:

`< interval begin >< interval end><gap><interval begin><interval end>`

This PR removes the gap intervals that could lead to mismatch between server stats and client stats.